// libs/my-library/src/enum/country.enum.ts
export enum CountryEnum {
  'NGN' = 'NGN',
  'USD' = 'USD',
  'GHS' = 'GHS',
  'OTHERS' = 'OTHERS',
  'GBP' = 'GBP',
  'MXN' = 'MXN',
  'ZMW' = 'ZMW',
  'ZWL' = 'ZWL',
  'KES' = 'KES',
  'TZS' = 'TZS',
  'ZAR' = 'ZAR',
  'UGX' = 'UGX',
  'RWF' = 'RWF',
}

export enum CountryCodeEnum {
  NGN = '+234',
  USA = '+1',
  GHS = '+233',
  MXN = '+52',
  GBP = '+44',
  ZMW = '+260',
  ZWL = '+263',
  KES = '+254',
  TZS = '+255',
  ZAR = '+27',
  UGX = '+256',
  RWF = '+250',
  // add other countries
}

export enum VestiLogosAndImages {
  vesti = 'https://res.cloudinary.com/wevesti/image/upload/v1618263060/WhatsApp_Image_2021-04-12_at_19.29.15_lh7uda.jpg',
}